Move the ncurses{w,}5-config scripts to their -dev packages
This requires removing the libdir variable whose value differs across
architectures.  Few packages seem to use the --libdir option, and none
should depend on its output since it is a directory which is already
in the linker's default search path.

There is no need for the -dev packages to require a specific
ncurses-bin version anymore, so drop the dependency.