Temporarily add a build-dependency on libncursesw5-dev
Sadly this is necessary since the package otherwise will FTBFTS with
ncurses-bin 5.9+20150516-1.  This build-dependency should be backed
out in a few weeks when it is made sure that no buildd has the broken
ncurses-bin version anymore.

A proper fix would be to use the pkg-config and ncursesw5-config files
from the build tree, but this needs more thought.