Install the wide library into a temporary scratchdir
This is necessary so that the .pc files are created and test/configure
can find them.

Unfortunately, configuring the test programs
"--with-curses-dir=$(wobjdir)\usr" to actually use the installed
version of the wide library does not work, since the configure script
assumes the library in the lib\ subdirectory and does not know about
multiarch, but using the library from the build tree is good enough.