Ship an empty /usr/share/terminfo directory in ncurses-base
Moving the *-config scripts to the -dev packages has triggered two
unexpected FTBFS bugs in jed and slrn which do not build-depend on
libncurses5-dev or libncursesw5-dev.

The configure script of these packages tries to run
"ncurses5-config --terminfo", and if this does not succeed, uses a
hardcoded list of terminfo directories which unfortunately does not
include the directories /etc/terminfo and /lib/terminfo.  Failing to
find a terminfo directory, it then concludes that the system must be
using termcap and adds "-ltermcap" to the linker line which fails.

Shipping /usr/share/terminfo works around this bug, since
/usr/share/terminfo is included in the hardcoded list of terminfo
directories.