Do not include the tic library in the libtinfo5-udeb package
It is not needed in the installer, and on architectures which build
multilib packages it has caused wrong dependencies for libtinfo5-udeb,
see https://bugs.debian.org/cgi-bin/bugreport.cgi?bug=821025.

To take care of the unlikely case that some udeb package includes a
binary linked to libtic, adjust the libtinfo5 shlibs file.