Enable parallel builds
Honor DEB_BUILD_OPTIONS=parallel=n, as per Policy §4.9.1.  This should
speed up building the package on the buildds quite a bit.

Since parallel builds will start several instances of the configure
script simultaneously, the output becomes rather incomprehensible by
default, therefore we use the -O option added in GNU Make 4.0.

As make 4.0 is already in stable, I did not bother to add a versioned
build dependency or a test for the feature.  It is not really likely
that somebody wants to backport ncurses to older releases anyway.