Invoke configure via relative paths
The full path to a source file showed up in various binaries, e.g. in
the ncurses-examples package, meaning that its content depended on
the build path.

According to
https://tests.reproducible-builds.org/debian/issues/unstable/captures_build_path_issue.html,
configure should be invoked via a relative and not absolute path, and
indeed this helps to produce identical binary packages even when the
build path is different.