Set LD_LIBRARY_PATH when building the test programs
The problem is that conftest will be linked to ncursesw, and produce
an error message if that library is not present on the system.  This
does not trigger right now because libncursesw5 is always installed,
but when we change the wide library to the new ABI it breaks.

./conftest: error while loading shared libraries: libncursesw.so.6: cannot open shared object file: No such file or directory

This might cause the build to fail later, depending on whether
libncursesw5-dev is installed.  In the latter case the fallback to
ncursesw5-config works, but otherwise the test programs will fail to
build.