Update libtinfo5.symbols file
Bring back the _nc_read_entry symbol, drop _nc_read_entry2.  This
restores ABI compatibility with the stretch libtinfo5 package, but
creates a new incompatibility with the version 6.0+20170708-1
currently in sid and testing, affecting partial upgrades from that
version.

To mitigate that problem, bump the minimal version for the
_nc_read_entry symbol to 6.0+20170715, ensuring that users who upgrade
ncurses-bin also get a newer libtinfo5 pulled in.

The other non-working partial upgrade attempt, upgrading libtinfo5 but
not ncurses-bin, does not seem to have a good solution.  Trying to use
a versioned "Breaks: ncurses-bin (= 6.0+20170708-1)" in libtinfo5
leads to a situation where there is no clean upgrade path and apt-get
croaks:

LANG=C apt-get install ../*20170715-1*.deb
[...]
E: This installation run will require temporarily removing the
essential package ncurses-bin:i386 due to a Conflicts/Pre-Depends
loop. This is often bad, but if you really want to do it, activate the
APT::Force-LoopBreak option.
E: Internal Error, Could not early remove ncurses-bin:i386 (2)

So I guess we will have to live with this problem and tell users who
run into it to upgrade the ncurses-bin package if necessary.