Add a lib64ncursesw6 package
I am not all keen to do this, but we are about to install the tinfo
libraries from the wide builds, and to do so for lib64tinfo6 the wide
build has to be enabled in the first place.  If that needs to be done,
we might just as well put the resulting 64-bit ncursesw library into a
package.  At least this establishes some symmetry between the 32- and
64-bit multilib builds.