Only compile the normal and wide libraries in an -indep build
Split up the build and install targets in debian/rules so that only
the configurations necessary to install and run tic are built and
installed in the -indep cases.  This includes the non-wide libraries,
because they must be installed before the wide ones as mentioned in
the INSTALL file.

It might be possible to reduce the build time for -arch only builds by
not running tic and skipping moving terminfo files around, but this
small gain is probably not worth it.  And running tic on all the
terminfo files also serves as a build-time test.