Build the static libraries without gpm support
For reasons unknown, the --disable-wattr-macros leads to a situation
where -lgpm is necessary to link any static program, even if it does
not use any mouse functions.  With the changes in the 20180609
patchlevel it would be possible to not configure the static libraries
separately, but then -ldl rather than -lgpm becomes required, and this
is rather undesirable in static binaries.

So disable gpm support in the static libraries altogether until
somebody figures out why --disable-wattr-macros causes this regression
and how to avoid it.