Export BUILD_{C,CPP,LD}FLAGS in debian/rules
When building the multilib packages, or during a cross build, two
helper programs in ncurses/tinfo (make_hash and make_keys) are built
with the host C compiler, which does not use {C,CPP,LD}FLAGS but
rather BUILD_{C,CPP,LD}FLAGS.  Not defining those meant that blhc
complained about missing flags.

Since the helper programs are not installed and only operate on known
input this is obviously harmless, and until now I had ignored the blhc
warnings, but now they trigger a failed pipeline in the Salsa CI
tests which is annoying.